Added missing dependency declarations to the pom.  Nixed mime-util.jar from project - we don't need it.  Code format fixes.

+++ src/main/java/com/tabulaw/server/AbstractDownloadServlet.java
@@ -2,7 +2,6 @@ package com.tabulaw.server;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.Collection;
 
 import javax.servlet.ServletException;
 import javax.servlet.ServletOutputStream;
@@ -18,15 +17,13 @@ import org.apache.commons.logging.LogFactory;
 import com.tabulaw.common.model.User;
 import com.tabulaw.service.convert.FileConverterDelegate;
 import com.tabulaw.util.StringUtil;
-import eu.medsea.mimeutil.MimeUtil;
 
 public abstract class AbstractDownloadServlet extends HttpServlet {
 
 	private static final long serialVersionUID = 842994896967308571L;
 	protected static final String EXPORT_TEMPLATE_PATH = "export-templates/";
 
-	protected static final Log log = LogFactory
-			.getLog(DocDownloadServlet.class);
+	protected static final Log log = LogFactory.getLog(DocDownloadServlet.class);
 
 	public static final String defaultMimeType = "text/html";
 
@@ -34,45 +31,37 @@ public abstract class AbstractDownloadServlet extends HttpServlet {
 	protected UserContext userContext;
 	protected String mimeType;
 
-	protected abstract File getContentFile(HttpServletRequest req)
-			throws ServletException, IOException;
+	protected abstract File getContentFile(HttpServletRequest req) throws ServletException, IOException;
 
-	private String getMimeType(File f) {
-		Collection<?> mimeTypes = MimeUtil.getMimeTypes(f);
-		return (mimeTypes.toString());
-
-	}
+//	private String getMimeType(File f) {
+//		Collection<?> mimeTypes = MimeUtil.getMimeTypes(f);
+//		return (mimeTypes.toString());
+//
+//	}
 
 	@Override
 	public void init() throws ServletException {
 		super.init();
-		MimeUtil.registerMimeDetector("eu.medsea.mimeutil.detector.MagicMimeMimeDetector");
+//		MimeUtil.registerMimeDetector("eu.medsea.mimeutil.detector.MagicMimeMimeDetector");
 	}
 
 	@Override
-	protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-			throws ServletException, IOException {
+	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 
 		HttpSession session = req.getSession(false);
-		if (session == null)
-			throw new ServletException("No user session");
+		if(session == null) throw new ServletException("No user session");
 
 		mimeType = req.getParameter("mimeType");
 
-		if (StringUtil.isEmpty(mimeType))
-			mimeType = defaultMimeType;
-		FileConverterDelegate fcd = (FileConverterDelegate) getServletContext()
-				.getAttribute(FileConverterDelegate.KEY);
+		if(StringUtil.isEmpty(mimeType)) mimeType = defaultMimeType;
+		FileConverterDelegate fcd = (FileConverterDelegate) getServletContext().getAttribute(FileConverterDelegate.KEY);
 
-		pc = (PersistContext) getServletContext().getAttribute(
-				PersistContext.KEY);
+		pc = (PersistContext) getServletContext().getAttribute(PersistContext.KEY);
 
 		userContext = (UserContext) session.getAttribute(UserContext.KEY);
-		if (userContext == null)
-			throw new ServletException("No user context in user session");
+		if(userContext == null) throw new ServletException("No user context in user session");
 		User user = userContext.getUser();
-		if (user == null)
-			throw new ServletException("No user in user context");
+		if(user == null) throw new ServletException("No user in user context");
 
 		File fdoc, fconverted;
 		fdoc = fconverted = null;
@@ -86,27 +75,24 @@ public abstract class AbstractDownloadServlet extends HttpServlet {
 			// stream the doc contents
 			ServletOutputStream sos = resp.getOutputStream();
 			byte[] fbytes = FileUtils.readFileToByteArray(fconverted);
-			if (fbytes == null || fbytes.length < 1)
-				throw new Exception("No doc content read");
-			//resp.setContentLength(fbytes.length);
-			resp.setContentType(getMimeType(fconverted));
-			resp.setHeader("Content-disposition", "attachment; filename="
-					+ fconverted.getName());
+			if(fbytes == null || fbytes.length < 1) throw new Exception("No doc content read");
+			// resp.setContentLength(fbytes.length);
+			//resp.setContentType(getMimeType(fconverted));
+			resp.setHeader("Content-disposition", "attachment; filename=" + fconverted.getName());
 			sos.write(fbytes);
 			responseWritten = true;
-		} catch (Exception e) {
+		}
+		catch(Exception e) {
 			String emsg = "Unable to convert:  due to error: " + e.getMessage();
-			if (!responseWritten)
-				resp.getWriter().write(emsg);
+			if(!responseWritten) resp.getWriter().write(emsg);
 			throw new ServletException(emsg, e);
-		} finally {
+		}
+		finally {
 			resp.flushBuffer();
 
 			// clean up
-			if (fdoc != null)
-				fdoc.delete();
-			if (fconverted != null)
-				fconverted.delete();
+			if(fdoc != null) fdoc.delete();
+			if(fconverted != null) fconverted.delete();
 		}
 	}
